added users who own their post or are an admin are the only ones who can delete a post

+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Post;
+use Illuminate\Support\Facades\Gate;
+
 
 class PostController extends Controller
 {
@@ -74,8 +76,14 @@ class PostController extends Controller
     public function destroy(string $id)
     {
         $post = Post::findOrFail($id);
-        $post->delete();
 
-        return redirect()->route('posts.index')->with('message', 'Post was deleted.');
+        $response = Gate::inspect('delete', $post);
+
+        if ($response->allowed()) {
+            $post->delete();
+            return redirect()->route('posts.index')->with('message', 'Post was deleted.');
+        } else {
+            return redirect()->route('posts.index')->with('message', $response->message());
+        }
     }
 }
